Edited elevstl.cpp to always produce STLs in the highest resolution possible
A C++ vector is now used instead of an array because of the resizing ability. Height and width variables are also set each time the program is run, possibly allowing for different size boxes in the future. The only potential problems with this new arrangement are the amount of memory the program will consume with this huge vector, and the size of the STLs produced. Solutions could include change hList to an int vector, and compressing the STLs before transfer to client.